Use named imports
Due to our plugin architecture, if this plugin is loaded first, it may
not include the code necessary to run `ChosenSelectInput`.

By using the named import, webpack will resolved the needed files and
the issue will be gone.

This is a temporary solution, we are currently rethinking our web plugin
architecture.